+++ Nginx/conf/nginx.conf
@@ -6,8 +6,11 @@ events {
 
 http {
     include       mime.types;
+    
     default_type  application/octet-stream;
+
     sendfile        on;
+
     keepalive_timeout  65;
 
     server {
@@ -30,7 +33,7 @@ http {
             proxy_set_header Upgrade $http_upgrade;
             proxy_set_header Connection $http_connection;
         }
-		
+
 		location ~* ^/(.*) {
 			client_max_body_size 10m;
