+++ RecipeBook.Web/lib/main.dart
@@ -19,7 +19,7 @@ class App extends StatelessWidget {
           Theme.of(context).textTheme,
         ),
       ),
-      home: const WidgetTree(),
+      home: WidgetTree(),
     );
   }
 }
