+++ RecipeBook.Web/lib/widget_tree.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:recipebook/home/home_page.dart';
-import 'package:recipebook/widgets/appbar_widget.dart';
+import 'package:recipebook/widgets/header_widget.dart';
 
 class WidgetTree extends StatelessWidget {
   const WidgetTree({Key? key}) : super(key: key);
@@ -8,11 +8,6 @@ class WidgetTree extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
-      extendBodyBehindAppBar: true,
-      appBar: PreferredSize(
-        preferredSize: Size(double.infinity, 78),
-        child: AppBarWidget(),
-      ),
       body: SingleChildScrollView(child: HomePage()),
     );
   }
